requirements-cuda.txt
# CUDA-specific requirements for Dia TTS Server
# This file is used when CUDA is detected

# PyTorch with CUDA support (CUDA 11.8 - widely compatible)
--index-url https://download.pytorch.org/whl/cu118
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
# torchvision not needed for TTS

# Core dependencies (same as requirements.txt)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
python-multipart>=0.0.6

# Audio processing
soundfile>=0.12.1
librosa>=0.10.1
numpy>=1.24.0
scipy>=1.10.0

# TTS Model
dia-tts

# Utils
rich>=13.0.0
huggingface-hub>=0.19.0
safetensors>=0.4.0
psutil>=5.9.0

# CUDA optimizations
triton>=2.1.0  # For torch.compile on Linux with CUDA
ninja  # For faster CUDA kernel compilation 